More tidying of render folder
