+++ app/blog/render/augment/list.js
